backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
Pillow==9.0.0
drf-extra-fields==3.6.1
isort==5.10.1
django-filter==23.2
django-import-export==3.2.0
flake8==4.0.1